Tpetra: refactor PerformanceCGSolve test
- Only run on one node at a time (so that Watchr output won't
  immediately get overwritten if Serial runs after Cuda)
- Add flags to control node type like MueLu does
  (e.g. in a Serial+Cuda build, default will be Cuda, and --serial and
  --cuda CLI flags control what is used)
- Combine everything into one cpp file. Don't need 4 files each
  instantiating for a specific node
