+++ packages/tpetra/core/test/PerformanceCGSolve/CMakeLists.txt
@@ -1,11 +1,4 @@
-
-SET(HEADERS "")
-SET(SOURCES "")
-
-
 SET_AND_INC_DIRS(DIR ${CMAKE_CURRENT_SOURCE_DIR})
-APPEND_GLOB(HEADERS ${DIR}/*.hpp)
-APPEND_GLOB(SOURCES ${DIR}/*.cpp)
 
 # The performance test currently depends on KokkosCore for its timers.
 # Otherwise, it does not use anything else in KokkosCore directly.
@@ -15,7 +8,7 @@ IF (Tpetra_INST_DOUBLE)
 
     # This test makes sense only for Scalar == double
     TRIBITS_ADD_EXECUTABLE( Performance-CGSolve
-                            SOURCES ${SOURCES}
+                            SOURCES cg_solve_file.cpp
                             COMM mpi
                             CATEGORIES BASIC PERFORMANCE
     )
